Clean up refactor
Use naming conventions consistently.
